+++ tests/test_guardian_persistence.py
@@ -18,7 +18,7 @@ from sqlalchemy.exc import IntegrityError, OperationalError
 pytestmark = pytest.mark.postgres
 
 
-HEAD_REVISION = "0032_decision_lifecycle_hardening"
+HEAD_REVISION = "0033_perspective_provenance"
 PREVIOUS_REVISION = "0006_portfolio_snapshot_status"
 
 GUARDIAN_TABLES = {